release/scripts/startup/model-maker/line_lib/requirements.txt
timm==0.6.12
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.5
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
matplotlib==3.4.2
sklearn-pandas==2.2.0
torchvision==0.10.0+cu111
tqdm==4.64.1
Pillow==8.2.0
typing-extensions==4.4.0
scikit-learn==0.24.2
